backend/requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
djangorestframework==3.12.4
environ==1.0
idna==2.10
PyJWT==1.7.1
python-http-client==3.3.2
pytz==2021.1
requests==2.25.1
sendgrid==6.7.1
six==1.16.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.1
twilio==6.60.0
urllib3==1.26.5
